build: split the tarball checks out of verify-pack
The documentation assertion took verify-pack.mts past the 300 line limit. What the tarball holds and what it weighs now live in verify-pack-tarball.mts, alongside the size ratchet they read.
